+++ packages/core/android/src/oldarch/java/io/sentry/react/RNSentryModule.java
@@ -152,6 +152,11 @@ public class RNSentryModule extends ReactContextBaseJavaModule {
     return this.impl.fetchNativePackageName();
   }
 
+  @ReactMethod
+  public void getDataFromUri(String uri, Promise promise) {
+    this.impl.getDataFromUri(uri, promise);
+  }
+
   @ReactMethod(isBlockingSynchronousMethod = true)
   public WritableMap fetchNativeStackFramesBy(ReadableArray instructionsAddr) {
     // Not used on Android